+++ BackEnd/src/main/java/com/ssafy/happyhouse/controller/FavoriteHouseController.java
@@ -5,6 +5,7 @@ import com.ssafy.happyhouse.domain.dto.FavoriteCountDTO;
 import com.ssafy.happyhouse.domain.dto.FavoriteRankDTO;
 import com.ssafy.happyhouse.domain.entity.HouseInfo;
 import com.ssafy.happyhouse.domain.entity.User;
+import com.ssafy.happyhouse.exception.AuthenticationRequiredException;
 import com.ssafy.happyhouse.service.FavoriteHouseService;
 import com.ssafy.happyhouse.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +21,6 @@ import java.util.List;
 public class FavoriteHouseController {
 
     private final FavoriteHouseService favoriteHouseService;
-    private final UserService userService;
 
     @GetMapping("/{aptCode}")
     public FavoriteCountDTO favoriteCount(@PathVariable Long aptCode) {
@@ -33,22 +33,38 @@ public class FavoriteHouseController {
 
     @PostMapping("/{aptCode}")
     public String addFavorite(@PathVariable Long aptCode, @Login User user) {
+        if(user == null){
+            log.warn("NO USER");
+            throw new AuthenticationRequiredException("로그인이 필요합니다.");
+        }
         favoriteHouseService.save(user.getId(), aptCode);
         return "ok";
     }
 
     @GetMapping("/users")
     public List<HouseInfo> findFavorite(@Login User user) {
+        if(user == null){
+            log.warn("NO USER");
+            throw new AuthenticationRequiredException("로그인이 필요합니다.");
+        }
         return favoriteHouseService.findFavoriteHouses(user.getId());
     }
 
     @GetMapping("/check/{aptCode}")
     public Boolean isFavorite(@PathVariable Long aptCode, @Login User user) {
+        if(user == null){
+            log.warn("NO USER");
+            throw new AuthenticationRequiredException("로그인이 필요합니다.");
+        }
         return favoriteHouseService.isFavorite(user.getId(), aptCode);
     }
 
     @DeleteMapping("/{aptCode}")
     public String deleteFavorite(@PathVariable Long aptCode, @Login User user) {
+        if(user == null){
+            log.warn("NO USER");
+            throw new AuthenticationRequiredException("로그인이 필요합니다.");
+        }
         favoriteHouseService.deleteById(aptCode, user);
         return "ok";
     }
